Update wp-show-posts.php
Fix the call to the_content (the first argument was incorrect). The default values will be used (null for the first argument and false for the second argument). With this change the more link will be generated. Not sure about what happens with the "read more" link text specified in the Column settings (if any). I verified that the change fixes the printing of the more link and now the "the_content_more_link" filter also works.